backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
pydantic==2.5.3
python-multipart==0.0.6
numpy==1.24.3
scipy==1.11.4
